+++ AdapterUnitTests/AppLovinAdapterTests/AUTAppLovinRewardedAdTests.m
@@ -52,6 +52,8 @@
   // Reset child-directed and under-age tags.
   GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = nil;
   GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = nil;
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentUnspecified;
 
   [super tearDown];
 }
@@ -114,6 +116,18 @@
 
 - (void)testLoadFailureIfUserIsTaggedAsUnderAge {
   GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = @YES;
+  AUTKMediationRewardedAdConfiguration *config =
+      [[AUTKMediationRewardedAdConfiguration alloc] init];
+
+  NSError *expectedError = [[NSError alloc] initWithDomain:GADMAdapterAppLovinErrorDomain
+                                                      code:GADMAdapterAppLovinErrorChildUser
+                                                  userInfo:nil];
+  AUTKWaitAndAssertLoadRewardedAdFailure(_adapter, config, expectedError);
+}
+
+- (void)testLoadFailureIfUserIsTaggedAsUnderAgeWithAgeRestrictedTreatment {
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentChild;
   AUTKMediationRewardedAdConfiguration *config =
       [[AUTKMediationRewardedAdConfiguration alloc] init];