+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobiInterstitialAd.m
@@ -118,10 +118,8 @@
   GADMAdapterInMobiRequestParametersMediationType mediationType =
       bidResponseData ? GADMAdapterInMobiRequestParametersMediationTypeRTB
                       : GADMAdapterInMobiRequestParametersMediationTypeWaterfall;
-  NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
-      extras, mediationType,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
+  NSDictionary<NSString *, id> *requestParameters =
+      GADMAdapterInMobiRequestParameters(extras, mediationType);
   [_interstitialAd setExtras:requestParameters];
 
   if (bidResponseData) {