+++ adapters/Unity/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Unity Ads iOS Mediation Adapter Changelog
 
+#### [Version 4.18.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/unity/UnityAdapter-4.18.0.0.zip)
+- Verified compatibility with Unity Ads SDK 4.18.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 13.3.0.
+- Unity Ads SDK version 4.18.0.
+
 #### [Version 4.17.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/unity/UnityAdapter-4.17.0.0.zip)
 - Verified compatibility with Unity Ads SDK 4.17.0.